Move validation to its own package
