+++ server/trace/mock_trace.go
@@ -15,12 +15,8 @@
 package trace
 
 import (
-	"fmt"
-	"reflect"
-	"strings"
-
-	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/googleinterns/cloud-operations-api-mock/validation"
 
 	"golang.org/x/net/context"
 
@@ -33,20 +29,9 @@ type MockTraceServer struct {
 	cloudtrace.UnimplementedTraceServiceServer
 }
 
-const (
-	invalidTimestampMsg   = "start time must be before end time"
-	malformedTimestampMsg = "unable to parse timestamp"
-	missingFieldMsg       = "span is missing required fields: %v"
-)
-
-var (
-	requiredMessageFields = []string{"DisplayName", "StartTime", "EndTime"}
-	requiredStringFields  = []string{"Name", "SpanId"}
-)
-
 func (s *MockTraceServer) BatchWriteSpans(ctx context.Context, req *cloudtrace.BatchWriteSpansRequest) (*empty.Empty, error) {
 	for _, span := range req.Spans {
-		if isValid, errMsg := isSpanValid(span); !isValid {
+		if isValid, errMsg := validation.IsSpanValid(span); !isValid {
 			return nil, status.Errorf(codes.InvalidArgument, errMsg)
 		}
 	}
@@ -54,55 +39,8 @@ func (s *MockTraceServer) BatchWriteSpans(ctx context.Context, req *cloudtrace.B
 }
 
 func (s *MockTraceServer) CreateSpan(ctx context.Context, span *cloudtrace.Span) (*cloudtrace.Span, error) {
-	if isValid, errMsg := isSpanValid(span); !isValid {
+	if isValid, errMsg := validation.IsSpanValid(span); !isValid {
 		return nil, status.Errorf(codes.InvalidArgument, errMsg)
 	}
 	return span, nil
-}
-
-func isSpanValid(span *cloudtrace.Span) (bool, string) {
-	if isValid, msg := validateRequiredFields(span); !isValid {
-		return isValid, msg
-	}
-
-	if isValid, msg := validateTimeStamps(span); !isValid {
-		return isValid, msg
-	}
-
-	return true, ""
-}
-
-func validateRequiredFields(span *cloudtrace.Span) (bool, string) {
-	var missingFields []string
-	spanReflect := reflect.ValueOf(span)
-
-	for _, field := range requiredStringFields {
-		fieldValue := reflect.Indirect(spanReflect).FieldByName(field).String()
-		if fieldValue == "" {
-			missingFields = append(missingFields, field)
-		}
-	}
-
-	for _, field := range requiredMessageFields {
-		fieldValue := reflect.Indirect(spanReflect).FieldByName(field)
-		if fieldValue.IsNil() {
-			missingFields = append(missingFields, field)
-		}
-	}
-
-	formattedErrMsg := fmt.Sprintf(missingFieldMsg, strings.Join(missingFields, ", "))
-	return len(missingFields) == 0, formattedErrMsg
-}
-
-func validateTimeStamps(span *cloudtrace.Span) (bool, string) {
-	start, err := ptypes.Timestamp(span.StartTime)
-	if err != nil {
-		return false, malformedTimestampMsg
-	}
-	end, err := ptypes.Timestamp(span.EndTime)
-	if err != nil {
-		return false, malformedTimestampMsg
-	}
-
-	return start.Before(end), invalidTimestampMsg
 }
